requirements.txt
# === CORE ===
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# === EMBEDDINGS ===
sentence-transformers==2.3.1
transformers==4.35.2

# === VECTOR DB ===
faiss-cpu==1.7.4

# === WEB API ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# === DATA ===
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4

# === UTILS ===
requests==2.31.0
httpx==0.26.0
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
orjson==3.9.10
psutil==5.9.6
colorlog==6.7.0

pytest==9.0.2
pytest-asyncio==0.23.6
pytest-cov==7.0.0
pytest-mock==3.14.0

# === LOGGING & MONITORING ===
prometheus-client==0.19.0